Added dropdowns to the top of the page to be used as filters.
Had to add ngxBootstrap 4 alpha-6 through npm for this to work properly on Angular
projects. If you're wondering what happened that caused this to take so
long, it took a long time for me to find that online, along with the computer updating to the Creators
Update in the middle of the day.