Updated the look and feel of the dropdowns.
1) The Dropdowns now activate on hover & on mouse click instead of just
on mouse click to better align with the rest of the website.
2) Dropdowns are now closer together in the middle and appear more like
their counterparts on the live site in the header (CSS taken directly
from the css on the live site).

Also added and removed a bunch of stuff relating to jQuery, but that
(hopefully) wont be shown as I'm pretty sure I uninstalled all/most of it.