Reverted filters to home. Rewrote to make better.
Reverted the filters to the home component from the filter-resources
service as it was not working properly in the service. Added a `TODO` to
check for why later.

Also rewrote the functions to allow the name of the dropdown to be
rewritten by only rewritting the string once and working everywhere.

Also added some more comments and did some miscellanious edits that
WebStorm recommended such as rewriting the if-statements for toggling
`resource.isVisible`.