Implement resources from REST onto resources page.
Almost all images are shown (those not do not have featured images in
Wordpress), all links (should) work and all titles are there.

The page takes a few seconds to load the first time (probably because of
the images).
TODO: Look into lazy loading for the page while getting the images.